fix: support fractional rollout percentages in local evaluation (#699)
* fix: support fractional rollout percentages

* fix: preserve rollout percentage API compatibility

* fix: use a double for rollout percentages

+++ posthog/src/main/java/com/posthog/internal/PostHogLocalEvaluationModels.kt
@@ -50,7 +50,7 @@ public class FlagFilters(
 public class FlagConditionGroup(
     public val properties: List<FlagProperty>?,
     @SerializedName("rollout_percentage")
-    public val rolloutPercentage: Int?,
+    public val rolloutPercentage: Double?,
     public val variant: String?,
 )
